+++ src/ocean/time/timeout/TimeoutManager.d
@@ -383,12 +383,12 @@ abstract class TimeoutManagerBase : ITimeoutManager
 
     ***************************************************************************/
 
-    public uint checkTimeouts ( bool delegate ( ITimeoutClient client ) dg = null )
+    public size_t checkTimeouts ( bool delegate ( ITimeoutClient client ) dg = null )
     {
         return this.checkTimeouts(this.now, dg);
     }
 
-    public uint checkTimeouts ( ulong now, bool delegate ( ITimeoutClient client ) dg = null )
+    public size_t checkTimeouts ( ulong now, bool delegate ( ITimeoutClient client ) dg = null )
     {
         debug ( TimeoutManager )
         {
